Add Keyboard::signal_activated

+++ simulant/generic/containers/polylist.h
@@ -116,7 +116,7 @@ public:
                  * so move to the next one that isn't empty */
                 auto next_chunk_id = current_->chunk + 1;
                 auto next_chunk = owner_->chunks_[next_chunk_id];
-                while((!next_chunk->used_list_head_) && next_chunk_id < owner_->chunks_.size() - 1) {
+                while((!next_chunk->used_list_head_) && next_chunk_id < (int) owner_->chunks_.size() - 1) {
                     next_chunk = owner_->chunks_[++next_chunk_id];
                 }
